Remove HostAPI from TODOs
The os-hosts API has been depraceted since microversion 2.43. The
suggested replacement is os-services API. The os-services API already
emits service.update notification for every action on that API. So we
agreed [1] not to transform the HostAPI notifications to the  versioned
notification format. So this patch removes them from our TODO list.

[1] http://lists.openstack.org/pipermail/openstack-dev/2017-September/122309.html